+++ lizard_wbconfiguration/tasks.py
@@ -23,7 +23,10 @@ from lizard_task.handler import get_handler
 def import_dbf(fews_meta_info=None,
                areas_filepath=None,
                buckets_filepath=None,
-               structures_filepath=None):
+               structures_filepath=None,
+               taskname="",
+               username=None,
+               levelno=20):
     """Import a waterbalance configuration from dbf.
 
     This function is provided for convenience only. It allows us to test the
@@ -31,12 +34,19 @@ def import_dbf(fews_meta_info=None,
     ConfigurationToValidate.
 
     """
+    handler = get_handler(taskname=taskname, username=username)
+    logger = logging.getLogger(taskname)
+    logger.addHandler(handler)
+    logger.setLevel(int(levelno))
+
     dbfimporter = DBFImporter()
     dbfimporter.fews_meta_info = fews_meta_info
     dbfimporter.areas_filepath = areas_filepath
     dbfimporter.buckets_filepath = buckets_filepath
     dbfimporter.structures_filepath = structures_filepath
     dbfimporter.import_dbf()
+
+    logger.removeHandler(handler)
     return "<<import dbf>>"
 
 def run_importdbf_task():
@@ -67,7 +77,7 @@ def validate_wbconfigurations(taskname="",
         config_type=configtype,
         action=action)
     v_configs = v_configs.exclude(file_path=None)
-    handler = get_handler(taskname, username)
+    handler = get_handler(taskname=taskname, username=username)
     logger = logging.getLogger(taskname)
     logger.addHandler(handler)
     logger.setLevel(int(levelno))
@@ -109,7 +119,7 @@ def validate_wbconfigurations(taskname="",
 
 
 @task()
-def validate_all():
+def validate_all(taskname='validate_all', username=None):
     """Import all currently available configurations.
 
     This method is a spike to see whether the import of water balance
@@ -124,7 +134,7 @@ def validate_all():
 
     """
     logger = logging.getLogger(__name__)
-    handler = get_handler('validate_all', 'admin')
+    handler = get_handler(taskname=taskname, username=username)
     logger.addHandler(handler)
     retriever = create_configurations_retriever()
     for configuration in retriever.retrieve_configurations():
@@ -156,7 +166,7 @@ def export_wbconfigurations_to_dbf(
     data_set -- name of organisation as DataSet in lizard_security
     levelno -- logging level as number, 10=debug, 20=info, ...
     """
-    handler = get_handler(taskname, username)
+    handler = get_handler(taskname=taskname, username=username)
     logger = logging.getLogger(taskname)
     logger.addHandler(handler)
     logger.setLevel(int(levelno))
@@ -186,16 +196,23 @@ def export_wbconfigurations_to_dbf(
 @task()
 def export_aanafvoergebieden_to_dbf(
     data_set=None,
-    taskname='aanafvoegebieden_export_to_dbf_all',
+    taskname='aanafvoergebieden_export_to_dbf_all',
     levelno=20,
     username=None):
     """
     Export geo info of 'aanafvoergebieden' into dbf.
     """
-    handler = get_handler(taskname, username)
+
+    handler = get_handler(taskname=taskname, username=username)
     logger = logging.getLogger(taskname)
     logger.addHandler(handler)
     logger.setLevel(int(levelno))
+
+    logger.info(data_set)
+    logger.info(taskname)
+    logger.info(username)
+    logger.info(levelno)
+
     dbfexporter = DBFExporter(logger)
     dbf_configurations = DBFConfiguration.objects.filter(dbf_type='Area')
     if data_set is not None:
@@ -208,6 +225,7 @@ def export_aanafvoergebieden_to_dbf(
         filename = dbf_configuration.filename
         if dbf_configuration.dbf_type == 'Area':
             dbfexporter.export_aanafvoergebieden(owner, save_to, filename)
+
     logger.info("Export water balance configurations is finished.")
     logger.removeHandler(handler)